refactor: EditForm and subscription management components
- Add EditForm component for editing car details with validation using react-hook-form and yup
- Introduce SubscriptionInfo, CheckoutButton, and ManageSubscriptionButton components for handling subscription actions
- Create CreateInvoiceForm for generating invoices with Stripe integration
- Update imports in Account and Invoices components to use new paths for the added components
- Add LabeledValue component for displaying labeled data
- Enhance GraphQL queries to include createdAt and updatedAt fields for cars
- Refactor pricing page to utilize new CheckoutButton component
