+++ src/components/common/EditForm/Subscription/SubscriptionInfo.tsx
@@ -3,8 +3,9 @@ import { Badge } from 'components/ui/badge'
 import { Button } from 'components/ui/button'
 import { Card, CardContent } from 'components/ui/card'
 import { Skeleton } from 'components/ui/skeleton'
-import { useSubscription } from '../../context/SubscriptionContext'
+
 import ManageSubscriptionButton from './ManageSubscriptionButton'
+import { useSubscription } from 'src/context/SubscriptionContext'
 
 interface SubscriptionInfoProps {
   showManageButton?: boolean
