+++ src/components/common/Invoicing/index.ts
@@ -1,0 +1,3 @@
+import CreateInvoiceForm from './CreateInvoiceForm'
+
+export default CreateInvoiceForm
