+++ src/components/common/LabeledValue/index.ts
